requirements.txt
Flask==3.0.0
stripe==7.11.0
python-dotenv==1.0.0
requests==2.32.3
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-escpos==2.2.0
